Renaming packages to old format (#130)
