+++ server/src/main/java/game/CrocodilePiece.java
@@ -1,6 +1,4 @@
-package game.pieces;
-
-import game.*;
+package Game;
 
 public class CrocodilePiece extends GamePiece {
     public CrocodilePiece(){
