+++ server/src/main/java/game/ElephantPiece.java
@@ -1,6 +1,4 @@
-package game.pieces;
-
-import game.*;
+package Game;
 
 public class ElephantPiece extends GamePiece {
