+++ server/src/main/java/game/LionPiece.java
@@ -1,6 +1,4 @@
-package game.pieces;
-
-import game.*;
+package Game;
 
 public class LionPiece extends GamePiece {
     /* initial constructor*/
