+++ server/src/main/java/game/PawnPiece.java
@@ -1,6 +1,4 @@
-package game.pieces;
-
-import game.*;
+package Game;
 
 public class PawnPiece extends GamePiece {
