+++ server/src/test/java/DatabaseHandlerTest.java
@@ -1,6 +1,6 @@
 import webconnection.Action;
 import database.DatabaseHandler;
-import game.Game;
+import Game.*;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.After;
