+++ server/src/test/java/MonkeyTest.java
@@ -1,6 +1,5 @@
-package game ;
+package Game ;
 
-import game.pieces.*;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
