+++ server/src/test/java/WebsocketServerTest.java
@@ -14,8 +14,7 @@ import org.junit.runners.JUnit4;
 import javax.annotation.processing.SupportedAnnotationTypes;
 
 import static org.junit.Assert.*;
-import game.pieces.*;
-import game.*;
+import Game.*;
 
 import java.net.InetSocketAddress;
 import java.nio.ByteBuffer;
